Outsource theme to improve readability and add fonts
