+++ src/react/App.jsx
@@ -4,32 +4,10 @@ import { createRoot } from 'react-dom/client'
 import { BrowserRouter as Router } from 'react-router-dom'
 
 import { CssBaseline } from '@mui/material'
-
-import { createTheme, ThemeProvider } from '@mui/material/styles'
+import { ThemeProvider } from '@mui/material/styles'
 
 import AppLayout from './AppLayout'
-
-const theme = createTheme({
-  components: {
-    MuiCssBaseline: {
-      styleOverrides: defaultTheme => ({
-        html: {
-          width: '100%',
-          height: '100%'
-        },
-        body: {
-          width: '100%',
-          height: '100%',
-          background: defaultTheme.palette.grey[200]
-        },
-        '#app': {
-          width: '100%',
-          height: '100%'
-        }
-      })
-    }
-  }
-})
+import theme from '../theme/theme'
 
 const container = document.getElementById('app')
 const root = createRoot(container)
